Added dashboard view which links all order admin pages
- change name 'catalog' to 'products'
- create dashboard_base template
- link settings page to dashboard and reformat
