--- inventory/templates/inventory/catalog.html
@@ -1,51 +1,0 @@
-{% extends 'base.html' %}
-
-{% block title %}Catalog{% endblock %}
-
-{% block content %}
-    <h1>Product Catalog</h1>
-    <div class="container">
-        <div class="row"><h3>Live Crop Products</h3></div>
-        <div class="row">
-            {% for lcp in live_crop_products %}
-                <div class="card" style="width: 18rem;">
-                    <div class="card-body">
-                        <h5 class="card-title">{{ lcp.name }}</h5>
-                        <h6 class="card-subtitle mb-2 text-muted">${{ lcp.price }}</h6>
-{#                        <p class="card-text">Some quick example text to build on the card title and make up the bulk of the card's content.</p>#}
-                        <a href="#" class="card-link">Edit product</a>
-                        <a href="#" class="card-link">Add similar product</a>
-                    </div>
-                </div>
-            {% endfor %}
-        </div>
-        <div class="row"><h3>Harvested Crop Products</h3></div>
-        <div class="row">
-            {% for hcp in harvested_crop_products %}
-                <div class="card" style="width: 18rem;">
-                    <div class="card-body">
-                        <h5 class="card-title">{{ hcp.name }}</h5>
-                        <h6 class="card-subtitle mb-2 text-muted">${{ hcp.price }}</h6>
-{#                        <p class="card-text">Some quick example text to build on the card title and make up the bulk of the card's content.</p>#}
-                        <a href="#" class="card-link">Edit product</a>
-                        <a href="#" class="card-link">Add similar product</a>
-                    </div>
-                </div>
-            {% endfor %}
-        </div>
-        <div class="row"><h3>Other Products</h3></div>
-        <div class="row">
-            {% for product in other_products %}
-                <div class="card" style="width: 18rem;">
-                    <div class="card-body">
-                        <h5 class="card-title">{{ product.name }}</h5>
-                        <h6 class="card-subtitle mb-2 text-muted">${{ product.price }}</h6>
-{#                        <p class="card-text">Some quick example text to build on the card title and make up the bulk of the card's content.</p>#}
-                        <a href="#" class="card-link">Edit product</a>
-                        <a href="#" class="card-link">Add similar product</a>
-                    </div>
-                </div>
-            {% endfor %}
-        </div>
-    </div>
-{% endblock %}
