+++ inventory/views.py
@@ -968,126 +968,3 @@ def add_barcodes(request):
                 slot.barcode = form.cleaned_data.pop('Slot ' + str(slot.id))
                 slot.save()
         return redirect(golden_trays_home)
-
-
-@staff_member_required
-def add_product(request):
-    if request.method == 'GET':
-        variety_list = []
-        for v in Variety.objects.all().order_by('name'):
-            variety_list.append(v.name)
-        sizes = []
-        for s in MicrogreenSize.objects.all().order_by('name'):
-            sizes.append(s.name)
-        tray_types = []
-        for t in TrayType.objects.all().order_by('name'):
-            tray_types.append(t.name)
-        product_form = AddProductForm()
-        return render(request, "inventory/add_product.html", context={"product_form": product_form,
-                                                                      "varieties": variety_list,
-                                                                      "tray_types": tray_types,
-                                                                      "sizes": sizes})
-    if request.method == 'POST':
-        variety_list = []
-        for v in Variety.objects.all().order_by('name'):
-            variety_list.append(v.name)
-        sizes = []
-        for s in MicrogreenSize.objects.all().order_by('name'):
-            sizes.append(s.name)
-        tray_types = []
-        for t in TrayType.objects.all().order_by('name'):
-            tray_types.append(t.name)
-        product_form = AddProductForm()
-        form = AddProductForm(request.POST)
-        if form.is_valid():
-            print(form.cleaned_data)
-            fields = request.POST
-            product_name = fields['product_name']
-            price = fields['price']
-            try:
-                if fields['select-product-type'] == 'Live Crop Product':
-                    print("Live crop product!")
-                    variety = Variety.objects.get(name=fields['variety'])
-                    size = MicrogreenSize.objects.get(name=fields['size'])
-                    tray_type = TrayType.objects.get(name=fields['tray-type'])
-                    LiveCropProduct.objects.create(name=product_name, price=price, variety=variety,
-                                                   size=size, tray_type=tray_type)
-                elif fields['select-product-type'] == 'Harvested Crop Product':
-                    variety = Variety.objects.get(name=fields['variety'])
-                    size = MicrogreenSize.objects.get(name=fields['size'])
-                    weight = int(fields['weight'])
-                    HarvestedCropProduct.objects.create(name=product_name, price=price, variety=variety,
-                                                        size=size, weight=weight)
-            except KeyError:
-                Product.objects.create(name=product_name, price=price)
-                pass
-            message = "Successfully added " + product_name + " to your "
-            # TODO: option to pre-fill previous form's values
-            return render(request, "inventory/add_product.html", context={"product_form": product_form,
-                                                                          "varieties": variety_list,
-                                                                          "tray_types": tray_types,
-                                                                          "sizes": sizes,
-                                                                          "success": message})
-        else:
-            message = "Something went wrong."
-            print(message)
-            return render(request, "inventory/add_product.html", context={"product_form": product_form,
-                                                                          "varieties": variety_list,
-                                                                          "tray_types": tray_types,
-                                                                          "sizes": sizes,
-                                                                          "error": message})
-
-
-@staff_member_required
-def catalog(request):
-    if request.method == 'GET':
-        other_products = []
-        for product in Product.objects.all():
-            if not LiveCropProduct.objects.filter(id=product.id).exists() \
-                    and not HarvestedCropProduct.objects.filter(id=product.id).exists():
-                other_products.append(product)
-        return render(request, "inventory/catalog.html", context={"live_crop_products": LiveCropProduct.objects.all(),
-                                                                  "harvested_crop_products": HarvestedCropProduct.objects.all(),
-                                                                  "other_products": other_products})
-    # if request.method == 'POST':
-    #     live_crop_products = []
-    #     harvested_crop_products = []
-    #     other_products = []
-    #     return render(request, "inventory/catalog.html", context={"live_crop_products": live_crop_products,
-    #                                                               "harvested_crop_products": harvested_crop_products,
-    #                                                               "other_products": other_products})
-
-
-@staff_member_required
-def orders_settings(request):
-    DAYS_OF_WEEK = (
-        (0, 'Sunday'),
-        (1, 'Monday'),
-        (2, 'Tuesday'),
-        (3, 'Wednesday'),
-        (4, 'Thursday'),
-        (5, 'Friday'),
-        (6, 'Saturday'),
-    )
-
-
-    if request.method == "GET":
-        delivery_days = eval(Setting.objects.get(name='DELIVERY_DAYS').value)
-        seeding_days = eval(Setting.objects.get(name='SEEDING_DAYS').value)
-        lead_time = eval(Setting.objects.get(name='CONSTANT_LEAD_DAYS').value)
-        return render(request, "inventory/orders_settings.html",
-                      context={"delivery_days": delivery_days, "seeding_days": seeding_days, "weekdays": DAYS_OF_WEEK, "lead_time": lead_time})
-    if request.method == "POST":
-        delivery_days = [int(i) for i in request.POST.getlist('delivery')]
-        seeding_days = [int(i) for i in request.POST.getlist('seeding')]
-        constant_lead = request.POST.get('constant-lead')
-        delivery_setting = Setting.objects.get(name='DELIVERY_DAYS')
-        seeding_setting = Setting.objects.get(name='SEEDING_DAYS')
-        constant_lead_setting = Setting.objects.get(name='CONSTANT_LEAD_DAYS')
-        delivery_setting.value = delivery_days
-        seeding_setting.value = seeding_days
-        constant_lead_setting.value = constant_lead if int(constant_lead) >= 0 else "0"
-        delivery_setting.save()
-        seeding_setting.save()
-        constant_lead_setting.save()
-        return redirect(orders_settings)
